pyproject.toml
[tool.poetry]
name = "portfolio_optimization_model"
version = "0.1.0"
description = ""
authors = ["RC", "LK", "SS"]
license = "PRIVATE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "~3.7.0"
yfinance = "^0.2.36"
scikit-learn = "^1.4.1"
python-dotenv = "^1.0.1"
pyportfolioopt = { version = "^1.5.5", python = ">=3.10,<3.15" }
nasdaq-data-link = "^1.0.4"
fredapi = "^0.5.1"
pyarrow = "^15.0.0"
quandl = "^3.7.0"
pydantic = "^2.6.1"
openpyxl = "^3.1.2"
seaborn = "^0.13.2"
scipy = "^1.12.0"

[tool.poetry.group.dev.dependencies]
cmake = "^3.28.3"
ruff = "^0.2.1"
pre-commit = "^3.6.1"
pytest = "^8.0.1"
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
